Add files via upload
I recently participated in Accenture’s Data Analytics & Visualization job simulation on the Forage platform, and it was incredibly useful to understand what it might be like to participate in a typical project team at Accenture, and to use Excel and data analytics and visualization skills in a realistic context. 
Through this program, I realized that I enjoy cleaning, modelling and analyzing client data, creating slides to communicate findings and presenting insights back to the client. I would love to apply what I've learned in a diverse project team at a company like Accenture.